more refactoring

+++ src/basic_ray_trace.rs
@@ -20,7 +20,7 @@ pub fn basic_ray_trace(filename: &Path, transform: Transform) -> Result<()> {
     let ray_origin = Point::point(0., 0., -5.);
     let wall_z = 10. as RayTracerFloat;
     let wall_size = 7. as RayTracerFloat;
-    let canvas_pixels = 100 as usize;
+    let canvas_pixels = 100_usize;
     let pixel_size = wall_size / (canvas_pixels as RayTracerFloat);
     let half = wall_size / 2.;
